[FEAT] article 공용 type 생성 | Create shared type for article TALEARNT-373
