+++ src/pages/articles/WriteArticle/core/writeArticle.api.ts
@@ -1,9 +1,7 @@
 import { postAPI } from "@utils/apiMethods";
 
-import {
-  matchArticleBodyType,
-  presignedURLBodyType
-} from "@pages/articles/WriteArticle/core/writeArticle.type";
+import { matchArticleBodyType } from "@pages/articles/articles.type";
+import { presignedURLBodyType } from "@pages/articles/WriteArticle/core/writeArticle.type";
 
 // TODO 커뮤니티용 body 가져오기
 export const postArticle = async (body: matchArticleBodyType) =>
