+++ src/pages/articles/WriteArticle/core/writeArticle.constants.ts
@@ -6,7 +6,7 @@ import {
   durationType,
   exchangeType,
   postType
-} from "@pages/articles/WriteArticle/core/writeArticle.type";
+} from "@pages/articles/articles.type";
 
 const articleTypeOptions = [
   { label: "매칭 게시물 글쓰기", value: "match" },
@@ -19,7 +19,7 @@ const durationList: durationType[] = [
   "3개월",
   "3개월 이상"
 ];
-const exchangeTypeList: exchangeType[] = ["온라인", "오프라인", "온오프라인"];
+const exchangeTypeList: exchangeType[] = ["온라인", "오프라인", "온_오프라인"];
 const postTypeList: postType[] = ["자유 게시판", "질문 게시판", "스터디 모집"];
 
 const durationOptions = durationList.map(item => ({
@@ -38,32 +38,14 @@ const receiveTalentsOptions = CATEGORIZED_TALENTS_LIST.map(
 
 const matchArticleSchema = object({
   giveTalents: array()
-    .of(
-      object({
-        label: string().required(),
-        value: number().required()
-      }).required()
-    )
+    .of(number().required())
     .min(1, "재능 키워드를 선택해 주세요")
     .required(),
   receiveTalents: array()
-    .of(
-      object({
-        label: string().required(),
-        value: number().required()
-      }).required()
-    )
+    .of(number().required())
     .min(1, "재능 키워드를 선택해 주세요")
     .required(),
-  duration: array()
-    .of(
-      object({
-        label: string().required(),
-        value: string().oneOf(durationList).required()
-      }).required()
-    )
-    .min(1, "진행 기간을 선택해 주세요")
-    .required(),
+  duration: string().oneOf(durationList).required(),
   exchangeType: string().oneOf(exchangeTypeList).required(),
   title: string()
     .required("제목을 입력해 주세요")
